switched to jersey, replacing esteasy added swagger ui & KeyPair / ChatApi partially
