+++ rest/src/main/java/bisq/api/jax/CustomExceptionMapper.java
@@ -1,0 +1,16 @@
+package bisq.api.jax;
+
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class CustomExceptionMapper implements ExceptionMapper<Exception> {
+    @Override
+    public Response toResponse(Exception exception) {
+        log.error(exception.getMessage(), exception);
+        return Response.status(Response.Status.BAD_REQUEST)
+                .entity(exception.getMessage())
+                .build();
+    }
+}
