+++ rest/src/main/java/bisq/api/jax/ProtoWriter.java
@@ -1,5 +1,7 @@
-package bisq.api.resteasy;
+package bisq.api.jax;
 
+import bisq.common.proto.Proto;
+import com.google.protobuf.util.JsonFormat;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.MediaType;
@@ -14,15 +16,15 @@ import java.lang.reflect.Type;
 
 @Provider
 @Produces(MediaType.APPLICATION_JSON)
-public class ProtoWriter implements MessageBodyWriter<Pet> {
+public class ProtoWriter implements MessageBodyWriter<Proto> {
 
     @Override
     public boolean isWriteable(Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType) {
-        return type == Pet.class;
+        return type.isAssignableFrom(Proto.class);
     }
 
     @Override
-    public void writeTo(Pet pet, Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> httpHeaders, OutputStream entityStream) throws IOException, WebApplicationException {
-        entityStream.write("{\"name\":\"pets are all equal\"}".getBytes());
+    public void writeTo(Proto proto, Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> httpHeaders, OutputStream entityStream) throws IOException, WebApplicationException {
+        entityStream.write(JsonFormat.printer().print(proto.toProto()).getBytes());
     }
 }
