+++ rest/src/main/resources/doc/v1/favicon-16x16.png
