+++ rest/src/main/resources/doc/v1/favicon-32x32.png
