+++ rest/src/main/resources/doc/v1/index.html
