+++ rest/src/main/resources/doc/v1/oauth2-redirect.html
