+++ rest/src/main/resources/doc/v1/swagger-initializer.js
@@ -3,7 +3,7 @@ window.onload = function() {
 
   // the following lines will be replaced by docker/configurator, when it runs in a docker-container
   window.ui = SwaggerUIBundle({
-    url: "/openapi.json",
+    url: "/rest/api/openapi.json",
     dom_id: '#swagger-ui',
     deepLinking: true,
     presets: [
