+++ rest/src/main/resources/doc/v1/swagger-ui-es-bundle-core.js
