+++ rest/src/main/resources/doc/v1/swagger-ui.css
